sailui+sidesail: move a bunch of code to sailui
