+++ packages/sail_ui/lib/bitcoin.dart
@@ -35,3 +35,11 @@ String formatDepositAddress(String address, int sidechainNum) {
 
   return strDepositAddress;
 }
+
+double cleanAmount(double amount) {
+  // clean the amount of any excess decimals! Cap it at 8, and let
+  // the front-end be kinda stupid
+  final cleanAmount = double.parse(amount.toStringAsFixed(8));
+
+  return cleanAmount;
+}
