+++ packages/sail_ui/lib/sail_ui.dart
@@ -6,6 +6,10 @@ export './widgets/buttons/toggle.dart';
 export './widgets/chips/action_header_chip.dart';
 export './widgets/chips/connection_error_chip.dart';
 export './widgets/chips/connection_success_chip.dart';
+export './widgets/components/core_transaction.dart';
+export './widgets/components/expanded_tx_view.dart';
+export './widgets/components/raw_transaction.dart';
+export './widgets/components/utxo.dart';
 export './widgets/containers/action_tile.dart';
 export './widgets/containers/notification.dart';
 export './widgets/containers/questions.dart';
