+++ packages/sail_ui/pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
     flutter_svg: ^2.0.10+1
     flutter_platform_alert: ^0.5.1
     intl: ^0.19.0
+    crypto: ^3.0.1
 
 dev_dependencies:
     flutter_test:
