+++ packages/sidesail/lib/pages/tabs/sidechain_explorer_tab_page.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/components/dashboard_group.dart';
 import 'package:sidesail/app.dart';
 import 'package:sidesail/config/dependencies.dart';
 import 'package:sidesail/config/sidechains.dart';
@@ -11,7 +12,6 @@ import 'package:sidesail/rpc/rpc_sidechain.dart';
 import 'package:sidesail/storage/client_settings.dart';
 import 'package:sidesail/storage/sail_settings/theme_settings.dart';
 import 'package:sidesail/widgets/containers/chain_overview_card.dart';
-import 'package:sidesail/widgets/containers/tabs/dashboard_tab_widgets.dart';
 import 'package:stacked/stacked.dart';
 
 @RoutePage()
