+++ packages/sidesail/lib/pages/tabs/sidechain_send_page.dart
@@ -3,17 +3,18 @@ import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
 import 'package:sail_ui/sail_ui.dart';
+import 'package:sail_ui/widgets/components/dashboard_group.dart';
 import 'package:sail_ui/widgets/core/sail_text.dart';
 import 'package:sidesail/config/sidechains.dart';
 import 'package:sidesail/providers/transactions_provider.dart';
 import 'package:sidesail/routing/router.dart';
-import 'package:sidesail/rpc/models/core_transaction.dart';
 import 'package:sidesail/rpc/rpc_sidechain.dart';
 import 'package:sidesail/widgets/containers/tabs/dashboard_tab_widgets.dart';
 import 'package:stacked/stacked.dart';
 
 @RoutePage()
 class SidechainSendPage extends StatelessWidget {
+  SidechainContainer get _sidechain => GetIt.I.get<SidechainContainer>();
   AppRouter get router => GetIt.I.get<AppRouter>();
 
   const SidechainSendPage({super.key});
@@ -63,6 +64,7 @@ class SidechainSendPage extends StatelessWidget {
                         for (final tx in viewModel.transactions)
                           CoreTransactionView(
                             tx: tx,
+                            ticker: _sidechain.rpc.chain.ticker,
                           ),
                       ],
                     ),
@@ -73,33 +75,6 @@ class SidechainSendPage extends StatelessWidget {
           ),
         );
       }),
-    );
-  }
-}
-
-class ExpandedTXView extends StatelessWidget {
-  final Map<String, dynamic> decodedTX;
-  final double width;
-
-  const ExpandedTXView({
-    super.key,
-    required this.decodedTX,
-    required this.width,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SailColumn(
-      spacing: SailStyleValues.padding08,
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: decodedTX.keys.where((key) => key != 'walletconflicts').map((key) {
-        return SingleValueContainer(
-          label: key,
-          value: decodedTX[key],
-          width: width,
-        );
-      }).toList(),
     );
   }
 }
