+++ packages/sidesail/lib/providers/cast_provider.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
-import 'package:sidesail/bitcoin.dart';
+import 'package:sail_ui/bitcoin.dart';
 import 'package:sidesail/providers/zcash_provider.dart';
 import 'package:sidesail/rpc/models/zcash_utxos.dart';
 import 'package:sidesail/rpc/rpc_zcash.dart';
