+++ packages/sidesail/lib/providers/transactions_provider.dart
@@ -3,8 +3,7 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:get_it/get_it.dart';
 import 'package:logger/logger.dart';
-import 'package:sidesail/rpc/models/core_transaction.dart';
-import 'package:sidesail/rpc/models/utxo.dart';
+import 'package:sail_ui/sail_ui.dart';
 import 'package:sidesail/rpc/rpc_mainchain.dart';
 import 'package:sidesail/rpc/rpc_sidechain.dart';
