+++ packages/sidesail/lib/widgets/containers/chain_overview_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:sail_ui/bitcoin.dart';
 import 'package:sail_ui/sail_ui.dart';
 import 'package:sail_ui/theme/theme.dart';
 import 'package:sail_ui/widgets/core/sail_text.dart';
-import 'package:sidesail/bitcoin.dart';
 import 'package:sidesail/config/sidechains.dart';
 
 class ChainOverviewCard extends StatelessWidget {
