+++ packages/sidesail/test/mocks/rpc_mock_mainchain.dart
@@ -1,7 +1,7 @@
+import 'package:sail_ui/sail_ui.dart';
 import 'package:sidesail/pages/tabs/settings/settings_tab.dart';
 import 'package:sidesail/rpc/models/active_sidechains.dart';
 import 'package:sidesail/rpc/models/blockchain_info.dart';
-import 'package:sidesail/rpc/models/utxo.dart';
 import 'package:sidesail/rpc/rpc_mainchain.dart';
 
 class MockMainchainRPC extends MainchainRPC {
