+++ packages/sidesail/test/mocks/rpc_mock_sidechain.dart
@@ -1,7 +1,7 @@
-import 'package:sidesail/bitcoin.dart';
+import 'package:sail_ui/bitcoin.dart';
+import 'package:sail_ui/sail_ui.dart';
 import 'package:sidesail/config/sidechains.dart';
 import 'package:sidesail/pages/tabs/settings/settings_tab.dart';
-import 'package:sidesail/rpc/models/core_transaction.dart';
 import 'package:sidesail/rpc/rpc_sidechain.dart';
 
 class MockSidechainRPC extends SidechainRPC {
